requirements.txt
streamlit==1.39.0
faiss-cpu==1.12.0
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.2.0
tqdm==4.66.5
PyMuPDF==1.24.9
numpy==1.26.4
pydantic==2.8.2
openai==1.50.2
